refactor(anthropic): consolidate cache checkpoint mechanism in cache_control.go
Pure move of extractSystemBlocks and extractBetaSystemBlocks so the
entire checkpoint mechanism (budget ledger, system-block marking with
cap, message-tail breakpoints) lives in one file, and document the
CacheControl contract on chat.Message: request-assembly state, set by
the session on assembled copies, JSON-serialized only because marks
must survive the before_llm_call hook rewrite round-trip.

Note: deleting the field in favor of out-of-band boundary indices was
considered and rejected \u2014 hooks.Input.Messages / updated_messages
round-trip assembled messages through external hook processes as JSON,
and per-message marks are the only representation that survives
hook-driven message-list rewrites (index-based boundaries would be
silently misplaced by an insert/delete).
